+++ packages/aside/index.ts
@@ -1,2 +1,2 @@
-import 'theme-chalk/src/aside.scss'
+import '../../static/theme-chalk/src/aside.scss'
 export { default } from './SeuAside.svelte'