+++ packages/button/index.ts
@@ -1,2 +1,2 @@
-import 'theme-chalk/src/button.scss'
+import '../../static/theme-chalk/src/button.scss'
 export { default } from './SeuButton.svelte'