+++ packages/checkbox/index.ts
@@ -1,2 +1,2 @@
-import 'theme-chalk/src/checkbox.scss'
+import '../../static/theme-chalk/src/checkbox.scss'
 export { default } from './SeuCheckbox.svelte'