+++ packages/col/index.ts
@@ -1,2 +1,2 @@
-import 'theme-chalk/src/col.scss'
+import '../../static/theme-chalk/src/col.scss'
 export { default } from './SeuCol.svelte'