+++ packages/header/index.ts
@@ -1,2 +1,2 @@
-import 'theme-chalk/src/header.scss'
+import '../../static/theme-chalk/src/header.scss'
 export { default } from './SeuHeader.svelte'