+++ packages/main/index.ts
@@ -1,2 +1,2 @@
-import 'theme-chalk/src/main.scss'
+import '../../static/theme-chalk/src/main.scss'
 export { default } from './SeuMain.svelte'