+++ packages/radio/index.ts
@@ -1,2 +1,2 @@
-import 'theme-chalk/src/radio.scss'
+import '../../static/theme-chalk/src/radio.scss'
 export { default } from './SeuRadio.svelte'