+++ packages/row/SeuRow.ts
@@ -1,2 +1,2 @@
-import 'theme-chalk/src/row.scss'
+import '../../static/theme-chalk/src/row.scss'
 export { default } from './SeuRow.svelte'