+++ src/main.ts
@@ -1,7 +1,7 @@
 import App from './App.svelte'
 import 'font-folder/element-icons.ttf'
 import 'font-folder/element-icons.woff'
-import 'theme-chalk/src/icon.scss'
+import '../static/theme-chalk/src/icon.scss'
 
 const app = new App({
   target: document.body,